refactor: `loader` → `loaders`
