+++ docs/src/content/docs/guides/site-search.mdx
@@ -129,7 +129,7 @@ Add translations of the modal UI for your language using Starlight’s built-in 
    ```js ins={5} ins=/{ extend: .+ }/
    // src/content.config.ts
    import { defineCollection } from 'astro:content';
-   import { docsLoader, i18nLoader } from '@astrojs/starlight/loader';
+   import { docsLoader, i18nLoader } from '@astrojs/starlight/loaders';
    import { docsSchema, i18nSchema } from '@astrojs/starlight/schema';
    import { docSearchI18nSchema } from '@astrojs/starlight-docsearch/schema';
