+++ examples/tailwind/src/content.config.ts
@@ -1,5 +1,5 @@
 import { defineCollection } from 'astro:content';
-import { docsLoader } from '@astrojs/starlight/loader';
+import { docsLoader } from '@astrojs/starlight/loaders';
 import { docsSchema } from '@astrojs/starlight/schema';
 
 export const collections = {
