+++ packages/starlight/routes/static/404.astro
@@ -2,7 +2,7 @@
 import { getEntry } from 'astro:content';
 import project from 'virtual:starlight/project-context';
 import config from 'virtual:starlight/user-config';
-import { getCollectionPathFromRoot } from '../../loader';
+import { getCollectionPathFromRoot } from '../../loaders';
 import {
 	normalizeCollectionEntry,
 	type Route,
